+++ src/main/java/com/devcoop/kiosk/domain/receipt/presentation/SaveReceiptController.java
@@ -6,6 +6,7 @@ import com.devcoop.kiosk.global.exception.GlobalException;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -20,8 +21,9 @@ public class SaveReceiptController {
 
   @PostMapping("/save/receipt")
   @Operation(summary = "save receipt", description = "키오스크에서 발행된 영수증 저장")
-  public void saveReceipt(@RequestBody KioskRequest kioskRequest) throws GlobalException {
-    System.out.println("영수증 발행");
+  public ResponseEntity<String> saveReceipt(@RequestBody KioskRequest kioskRequest) throws GlobalException {
     receiptService.saveReceipt(kioskRequest);
+    System.out.println("영수증 발행");
+    return ResponseEntity.ok("영수증 발행 성공");
   }
 }
