+++ src/main/java/com/devcoop/kiosk/domain/user/service/UserAuthService.java
@@ -4,15 +4,12 @@ import com.devcoop.kiosk.domain.user.User;
 import com.devcoop.kiosk.domain.user.presentation.dto.LoginRequest;
 import com.devcoop.kiosk.domain.user.presentation.dto.LoginResponse;
 import com.devcoop.kiosk.domain.user.presentation.dto.PinChangeRequest;
-import com.devcoop.kiosk.domain.user.utils.JwtUtil;
+import com.devcoop.kiosk.global.utils.security.JwtUtil;
 import com.devcoop.kiosk.domain.user.repository.UserRepository;
 import com.devcoop.kiosk.global.exception.GlobalException;
 import com.devcoop.kiosk.global.exception.enums.ErrorCode;
 import lombok.RequiredArgsConstructor;
-import org.apache.coyote.Response;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
